+++ visual_rag/retrieval/three_stage.py
@@ -51,7 +51,7 @@ class ThreeStageRetriever:
         if isinstance(embedding, torch.Tensor):
             if embedding.dtype == torch.bfloat16:
                 return embedding.cpu().float().numpy()
-            return embedding.cpu().numpy()
+            return embedding.cpu().float().numpy()  # .float() for BFloat16 compatibility
         return np.array(embedding, dtype=np.float32)
 
     def _infer_vector_is_multivector(self, vector_name: str) -> bool:
